Refactored-Blackjack Complete
Refactored getHandTotal function to take in specific cards instead of the whole array. This is two card params after the initial deal with the second param set to default 0 and a break statement when the player or dealer is just hitting one card at a time.
Also added function to properly clear each hand of cards after the round. Blackjack is feature complete with maybe some light line spacing needed for visual  appreciation.