Roulette features add
player can choose a number to bet on, then choose an amount of money to bet. If they still have money in the bank, they have the option to choose and bet on another number and so on. A random number is then generated to use as an index number on the 'wheel' of roulette numbers. The players choices are then checked against the winning number and bank account is updated accordingly. If the bank is empty the player loses and the game ends. If the player still has money then they will be able to choose to play another round and all their previously chosen numbers and bets will be cleared.